Fix pr with feedback
Abstract tab in another component

Abstract tab in another component

Abstract tab in another component

Load devices dynamically

Abstract cpx

Update snapshot
